Make compatible with Python 3 and make work
The Quick2Wire library silently fails unless used with Python 3 (it does
say this in the documentation, but if you accidentally use Python 2.x
it will appear to work---no errors---but doesn't).  So, switched to
Python 3 and am now able to control the LED strip through SPI.

Reread the Adafruit documentation on the wire protocol and changed the
code so that it now works. Necessary to send a latch byte (0) after the
36 bytes of color data to have the last byte get latched (the blue of
the final LED) and also send a zero at the start to reset the strip.

This version of the code works and downloads the weather from the BBC
and updates the LED strip. Further work:

  1. Choose good colors for the LEDs for each weather condition

  2. Look at missing weather conditions (such as hail)